+++ genset/controller_test.go
@@ -375,7 +375,7 @@ func TestController3P(t *testing.T) {
 
 			stateTracker.Assert(t, []genset.State{
 				{Node: genset.Cranking, Changed: t0},
-				{Node: genset.Error, Changed: t0.Add(params.CrankingTimeout)},
+				{Node: genset.Error, Changed: t0.Add(params.CrankingTimeout), ErrorTrigger: genset.ErrCrankingTimeout},
 			})
 
 			{ // check starter off
@@ -626,7 +626,7 @@ func TestController3P(t *testing.T) {
 
 			stateTracker.Assert(t, []genset.State{
 				{Node: genset.Producing, Changed: t0},
-				{Node: genset.Error, Changed: t1},
+				{Node: genset.Error, Changed: t1, ErrorTrigger: genset.ErrFrequencyHigh},
 			})
 
 			// check ignition is off
@@ -649,7 +649,7 @@ func TestController3P(t *testing.T) {
 
 			stateTracker.Assert(t, []genset.State{
 				{Node: genset.Producing, Changed: t0},
-				{Node: genset.Error, Changed: t1},
+				{Node: genset.Error, Changed: t1, ErrorTrigger: genset.ErrIOUnavailable},
 			})
 
 			// check ignition is off
@@ -1017,7 +1017,7 @@ func TestController1P(t *testing.T) {
 
 			stateTracker.Assert(t, []genset.State{
 				{Node: genset.Producing, Changed: t0},
-				{Node: genset.Error, Changed: t1},
+				{Node: genset.Error, Changed: t1, ErrorTrigger: genset.ErrU1High},
 			})
 
 			// check ignition is off